Users/rajes/OneDrive/Desktop/claude 2/AutoFilterBot/requirements.txt
pyrofork==2.3.38
tgcrypto==1.2.5
motor==3.4.0
aiohttp==3.10.5
python-dotenv==1.0.1
